pyproject.toml
[project]
name = "meeting-notes-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
  "fastapi>=0.128.0",
  "pydantic>=2.12.5",
  "python-dotenv>=1.2.1",
  "uvicorn>=0.40.0",
  "jinja2>=3.1.4",
  "python-multipart>=0.0.12",
  "openai>=2.14.0",
  "azure-ai-formrecognizer>=3.3.3",
  "pypdf2>=3.0.1",
]

   
    

